+++ src/gtest/test_keys.cpp
@@ -1,7 +1,6 @@
 #include <chainparams.h>
 #include <key_io.h>
 #include <zcash/Address.hpp>
-#include <zcash/zip32.h>
 
 #include "utiltest.h"
