+++ src/gtest/test_zip32.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
 
-#include <zcash/zip32.h>
+#include <zcash/address/zip32.h>
 
 // From https://github.com/zcash-hackworks/zcash-test-vectors/blob/master/sapling_zip32.py
 // Sapling consistently uses little-endian encoding, but uint256S takes its input in
