+++ src/key_io.h
@@ -12,7 +12,6 @@
 #include <pubkey.h>
 #include <script/standard.h>
 #include <zcash/Address.hpp>
-#include <zcash/zip32.h>
 
 #include <string>
