+++ src/wallet/asyncrpcoperation_saplingmigration.h
@@ -2,7 +2,6 @@
 #include "asyncrpcoperation.h"
 #include "univalue.h"
 #include "zcash/Address.hpp"
-#include "zcash/zip32.h"
 
 class AsyncRPCOperation_saplingmigration : public AsyncRPCOperation
 {
